Move create_db_instance to top test config
Why:

It is necessary in functional tests as well
